# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ConclusionResponse" : {
        "properties" : {
          "confidence" : {
            "format" : "double",
            "type" : "number"
          },
          "content" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "lastReinforcedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "level" : {
            "type" : "string"
          },
          "messageIds" : {
            "items" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "array"
          },
          "observed" : {
            "type" : "string"
          },
          "observer" : {
            "type" : "string"
          },
          "session" : {
            "type" : "string"
          },
          "sourceIds" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "timesDerived" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "ExplainResponse" : {
        "properties" : {
          "ent" : {
            "format" : "double",
            "type" : "number"
          },
          "kw" : {
            "format" : "double",
            "type" : "number"
          },
          "lvl" : {
            "format" : "double",
            "type" : "number"
          },
          "matchedEntities" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "rec" : {
            "format" : "double",
            "type" : "number"
          },
          "reinf" : {
            "format" : "double",
            "type" : "number"
          },
          "sem" : {
            "format" : "double",
            "type" : "number"
          },
          "weights" : {
            "items" : {
              "format" : "double",
              "type" : "number"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "MessageResponse" : {
        "properties" : {
          "content" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "metadata" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "peer" : {
            "type" : "string"
          },
          "seq" : {
            "format" : "int64",
            "type" : "integer"
          },
          "tokenCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ProvenanceEntry" : {
        "properties" : {
          "conclusion" : {
            "$ref" : "#/components/schemas/ConclusionResponse"
          },
          "premises" : {
            "items" : {
              "$ref" : "#/components/schemas/ConclusionResponse"
            },
            "type" : "array"
          },
          "sourceMessages" : {
            "items" : {
              "$ref" : "#/components/schemas/MessageResponse"
            },
            "type" : "array"
          },
          "unresolvedPremiseIds" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "ProvenanceResponse" : {
        "properties" : {
          "conclusions" : {
            "items" : {
              "$ref" : "#/components/schemas/ProvenanceEntry"
            },
            "type" : "array"
          },
          "entity" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "RecallHitResponse" : {
        "properties" : {
          "conclusion" : {
            "$ref" : "#/components/schemas/ConclusionResponse"
          },
          "explain" : {
            "$ref" : "#/components/schemas/ExplainResponse"
          },
          "score" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "RecallQuery" : {
        "properties" : {
          "explain" : {
            "type" : "boolean"
          },
          "filter" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "limit" : {
            "format" : "int32",
            "type" : "integer"
          },
          "observed" : {
            "minLength" : 1,
            "type" : "string"
          },
          "observer" : {
            "minLength" : 1,
            "type" : "string"
          },
          "query" : {
            "minLength" : 1,
            "type" : "string"
          },
          "threshold" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "required" : [ "observed", "observer", "query" ],
        "type" : "object"
      },
      "RecallResponseBody" : {
        "properties" : {
          "analyzedQuery" : {
            "type" : "string"
          },
          "candidatesConsidered" : {
            "format" : "int32",
            "type" : "integer"
          },
          "hits" : {
            "items" : {
              "$ref" : "#/components/schemas/RecallHitResponse"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "description" : "A token from `POST /v1/tokens`, in four nesting scopes: admin, workspace, peer, session. Each operation below names the narrowest one that may call it.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Pair-scoped memory. Every fact belongs to a directed (observer, observed) pair: what bob remembers about alice is not what alice remembers about herself, and no route crosses that line.\n\nThree tiers of read: `context` assembles a session's summary and recent messages within a token budget, `recall` ranks conclusions by six fused signals, and `chat` runs an agentic loop over both when the answer needs more than one lookup.",
    "license" : {
      "name" : "Apache-2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "aimon-memory",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/v1/workspaces/{workspace}/recall" : {
      "post" : {
        "description" : "**Requires a `peer` token or wider.**",
        "operationId" : "recall",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecallQuery"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecallResponseBody"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The filter parsed but was rejected: an unfilterable column, or an operand of the wrong type. Refused rather than quietly dropped, which would answer 200 with rows nobody asked for."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          }
        },
        "summary" : "Tier 1: rank a pair's conclusions against a query",
        "tags" : [ "recall" ]
      }
    },
    "/v1/workspaces/{workspace}/recall/provenance" : {
      "get" : {
        "description" : "**Requires a `peer` token or wider.**",
        "operationId" : "provenance",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "entity",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "observer",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "observed",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 10,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProvenanceResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          }
        },
        "summary" : "Trace an entity to the beliefs and sentences behind it",
        "tags" : [ "recall" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "description" : "Relative to wherever this service is deployed.",
    "url" : "/"
  } ],
...
}